extract StreamEffectAsyncSpec
